# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blossom Buddy API",
        "description": "API pour la gestion des plantes et utilisateurs",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/plant": {
            "get": {
                "tags": [
                    "Plants"
                ],
                "summary": "Récupérer toutes les plantes",
                "operationId": "7c23a6e1579ac9f5885edf17b4eddace",
                "responses": {
                    "200": {
                        "description": "Liste de toutes les plantes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "common_name": {
                                                "type": "string",
                                                "example": "Ficus"
                                            },
                                            "watering_general_benchmark": {
                                                "properties": {
                                                    "value": {
                                                        "type": "string",
                                                        "example": "5-7"
                                                    },
                                                    "unit": {
                                                        "type": "string",
                                                        "example": "days"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Plants"
                ],
                "summary": "Ajouter une nouvelle plante",
                "operationId": "bdf6e893a6c68eac93ffa14b741ea500",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "common_name",
                                    "watering_general_benchmark"
                                ],
                                "properties": {
                                    "common_name": {
                                        "type": "string",
                                        "example": "Ficus"
                                    },
                                    "watering_general_benchmark": {
                                        "properties": {
                                            "value": {
                                                "type": "string",
                                                "example": "5-7"
                                            },
                                            "unit": {
                                                "type": "string",
                                                "example": "days"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plante créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "common_name": {
                                            "type": "string",
                                            "example": "Ficus"
                                        },
                                        "watering_general_benchmark": {
                                            "type": "object"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/plant/{name}": {
            "get": {
                "tags": [
                    "Plants"
                ],
                "summary": "Récupérer une plante par son nom",
                "operationId": "7689b9b14a12aa70fcbe135e7f73bc32",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Nom commun de la plante",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Ficus"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informations de la plante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "common_name": {
                                            "type": "string",
                                            "example": "Ficus"
                                        },
                                        "watering_general_benchmark": {
                                            "type": "object"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plante non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Plante non trouvée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/plant/{id}": {
            "delete": {
                "tags": [
                    "Plants"
                ],
                "summary": "Supprimer une plante",
                "operationId": "533f3fb60e61315d56125d3e1efeb83b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la plante à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plante supprimée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Plante supprimée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plante non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Plante non trouvée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}